+++ examples/buongiornissimo.rs
@@ -1,4 +1,4 @@
-use buongiornissimo_rs::{BuongiornissimoCaffe, Scrape};
+use buongiornissimo_rs::{BuongiornissimoCaffe, BuongiornoImmagini, Scrape, TiCondivido};
 use chrono::Local;
 use rand::Rng;
 
@@ -8,8 +8,10 @@ fn choice<T>(choices: &[T]) -> &T {
 }
 
 fn get_provider() -> Box<dyn Scrape> {
-    match *choice(&[0]) {
+    match *choice(&[0, 1, 2]) {
         0 => Box::new(BuongiornissimoCaffe::default()),
+        1 => Box::new(BuongiornoImmagini::default()),
+        2 => Box::new(TiCondivido::default()),
         _ => panic!("out of range"),
     }
 }