+++ src/providers.rs
@@ -6,6 +6,8 @@ use super::{Greeting, Scrape, ScrapeError, ScrapeResult, Url};
 
 mod buongiornissimo_caffe;
 mod buongiorno_immagini;
+mod ticondivido;
 
 pub use self::buongiornissimo_caffe::BuongiornissimoCaffe;
 pub use self::buongiorno_immagini::BuongiornoImmagini;
+pub use self::ticondivido::TiCondivido;